Refactor properties
