+++ iam-login-service/src/main/java/it/infn/mw/iam/core/oauth/scope/pdp/DefaultScopeFilter.java
@@ -57,7 +57,7 @@ public class DefaultScopeFilter implements ScopeFilter {
       String clientId) {
     Optional<IamAccount> account = accountUtils.getAuthenticatedUserAccount(authn);
     if (account.isEmpty()) {
-      if (properties.getOpa().isEnabled()) {
+      if (properties.getScopeAuthz().getOpa().isEnabled()) {
         return filterScopes(requestedScopes, (IamAccount) null, clientId);
       }
       return requestedScopes;
