+++ iam-login-service/src/main/resources/application.yml
@@ -97,10 +97,6 @@ iam:
   topbarTitle: ${IAM_TOPBAR_TITLE:INDIGO IAM for ${iam.organisation.name}}
   showSql: ${IAM_DATABASE_SHOW_SQL:false}
 
-  opa:
-    enabled: ${IAM_OPA_ENABLED:false}
-    url: ${IAM_OPA_URL:http://localhost:8181/}
-
   jwk:
     keystore-location: ${IAM_KEY_STORE_LOCATION:}
     default-key-id: ${IAM_JWK_DEFAULT_KEY_ID:rsa1}
@@ -139,6 +135,9 @@ iam:
   scope-authz:
     enabled: ${IAM_ENABLE_SCOPE_AUTHZ:true}
     early-fail: ${IAM_EARLY_FAIL_SCOPE_AUTHZ:false}
+    opa:
+      enabled: ${IAM_OPA_ENABLED:false}
+      url: ${IAM_OPA_URL:http://localhost:8181/}
 
   access_token:
     include_authn_info: ${IAM_ACCESS_TOKEN_INCLUDE_AUTHN_INFO:false}
