+++ iam-login-service/src/test/java/it/infn/mw/iam/test/oauth/scope/pdp/ScopePolicyEarlyFailTests.java
@@ -76,7 +76,7 @@ class ScopePolicyEarlyFailTests extends ScopePolicyTestUtils {
   void tesScopesAreAllowed() {
 
     IamAccount testAccount = accountRepo.findByUsername("test")
-      .orElseThrow(() -> new AssertionError("Expected test account not found!"));;
+      .orElseThrow(() -> new AssertionError("Expected test account not found!"));
 
     IamScopePolicy up = initDenyScopePolicy();
     up.linkAccount(testAccount);
